% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIBT.R
\name{getIBT}
\alias{getIBT}
\title{getIBT}
\usage{
getIBT(
  intensityData,
  subjectName = "Panéliste",
  scoreName = "Score",
  productName = "Produit",
  decreasingConcentrations = NULL,
  decreasingNumConcentrations = NULL,
  minConc = NULL,
  maxConc = NULL
)
}
\arguments{
\item{intensityData}{data containing columns for subjectName (containing the name of the subject),scoreName (containing the intensity) and productName (containing the name of the concentration)}

\item{subjectName}{name of the column containing the subject information}

\item{scoreName}{name of the column containing the intensities}

\item{productName}{name of the column containing the subject information}

\item{decreasingConcentrations}{vector containing the names of the concentrations corresponding to the decreasing concentrations}

\item{decreasingNumConcentrations}{vector containing the numerical concentrations (logged or not)}

\item{minConc}{a number indicating the minimal concentration (not tested, but used for threshold calculations)}

\item{maxConc}{a number indicating the maximal concentration (not tested, but used for threshold calculations)}
}
\description{
A threshold is calculated as the start of the highest slope (cumulative while no negative slope)
}
