% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIntensityThreshold.R
\name{getIntensityThreshold}
\alias{getIntensityThreshold}
\title{getIntensityThreshold}
\usage{
getIntensityThreshold(
  intensityData,
  subjectName = "Panéliste",
  scoreName = "Score",
  productName = "Produit",
  decreasingConcentrations = NULL,
  decreasingNumConcentrations = NULL,
  minConc = NULL,
  maxConc = NULL,
  tol = 0,
  scaleBySubject = TRUE
)
}
\arguments{
\item{intensityData}{data containing columns for subjectName (containing the name of the subject),scoreName (containing the intensity) and productName (containing the name of the concentration)}

\item{subjectName}{name of the column containing the subject information}

\item{scoreName}{name of the column containing the intensities}

\item{productName}{name of the column containing the subject information}

\item{decreasingConcentrations}{vector containing the names of the concentrations corresponding to the decreasing concentrations}

\item{decreasingNumConcentrations}{vector containing the numerical concentrations (logged or not)}

\item{minConc}{a number indicating the minimal concentration (not tested, but used for threshold calculations)}

\item{maxConc}{a number indicating the maximal concentration (not tested, but used for threshold calculations)}

\item{tol}{O by default. If scaleBySubject=TRUE, the tolerance is tol. If scaleBySubject=TRUE, the tolerance is tol*(max(score_of_this_subject)-min(score(score_of_this_subject))}

\item{scaleBySubject}{if TRUE the tolerance is calculated by subject (as a proportion of the distance max-min of the scale)}
}
\description{
This function calculates an intensity threshold based on the first moment when the scores are only increasing in the data (in increasing concentrations). An increasing is defined up to a tolerance. A tolerance of 0 returns the moment when the is (not strictly increasing), a tolerance of 1 allows a small decreasing of 1.
}
