% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatrixTsachantS.r
\name{getMatrixTsachantS}
\alias{getMatrixTsachantS}
\title{getMatrixTsachantS}
\usage{
getMatrixTsachantS(p, K, eps = 0, modelEps = "last")
}
\arguments{
\item{p}{probability to fail an AFC test}

\item{K}{number of successive AFC tests}

\item{eps}{0}

\item{modelEps}{"last"}
}
\description{
getMatrixTsachantS
}
