% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScreeThreshold.R
\name{getScreeThreshold}
\alias{getScreeThreshold}
\title{getScreeThreshold}
\usage{
getScreeThreshold(
  intensityData,
  subjectName = "Panéliste",
  scoreName = "Score",
  productName = "Produit",
  decreasingConcentrations = NULL,
  decreasingNumConcentrations = NULL,
  minConc = NULL,
  maxConc = NULL
)
}
\arguments{
\item{subjectName}{name of the column containing the subject information}

\item{scoreName}{name of the column containing the intensities}

\item{productName}{name of the column containing the subject information}

\item{decreasingConcentrations}{vector containing the names of the concentrations corresponding to the decreasing concentrations}

\item{decreasingNumConcentrations}{vector containing the numerical concentrations (logged or not)}

\item{minConc}{a number indicating the minimal concentration (not tested, but used for threshold calculations)}

\item{maxConc}{a number indicating the maximal concentration (not tested, but used for threshold calculations)}
}
\description{
The threshold is calculated here as the maximale differences between two consecutive slopes
}
