% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getThreshold.r
\name{getThreshold}
\alias{getThreshold}
\title{returns the threshold (=0 if the subject found the right samples for all concentrations, 1 if he/she found the right samples for all concentrations but the lightest one, etc.}
\usage{
getThreshold(
  df,
  decreasingConcentrations = c("C9", "C8", "C7", "C6", "C5", "C4", "C3", "C2", "C1")
)
}
\arguments{
\item{df}{result of keepLastOccurrence}

\item{decreasingConcentrations}{vector contaning the products corresponding to the decreasing concentrations}
}
\description{
returns the threshold (=0 if the subject found the right samples for all concentrations, 1 if he/she found the right samples for all concentrations but the lightest one, etc.
}
