% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupBET.R
\name{groupBET}
\alias{groupBET}
\title{groupBET}
\usage{
groupBET(dfT, nb_total = 193, type = "geom")
}
\arguments{
\item{dfT}{dataframe whose colnames are "t" (containing the proportion of subject having this BET) and "BET", value of the BET}

\item{nb_total}{number of subjects in the distribution}

\item{type}{'geom' or 'arithm' for geometric or arithmetic average}
}
\description{
returns the BET of a group of subjects
}
