% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/houghCorrection.r
\name{houghCorrection}
\alias{houghCorrection}
\title{houghCorection}
\usage{
houghCorrection(triangularMatrix, p_chance = 1/3, numericConcentrations = NULL)
}
\arguments{
\item{triangularMatrix}{a matrix containing in row the subjects, in line the results for each concentration (decreasing). The score is 1 if success, 0 if failure}

\item{p_chance}{probability to give the good answer by chance}

\item{numericConcentrations}{increasing numericConcentrations}
}
\description{
returns a Hough-corrected matrix from triangular matrix
}
