% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keepLastOccurrence.r
\name{keepLastOccurence}
\alias{keepLastOccurence}
\title{keepLastOccurence}
\usage{
keepLastOccurence(
  x,
  subjectName = "Panéliste",
  productName = "Produit",
  descriptorName = "Descripteur",
  timeName = "Temps"
)
}
\arguments{
\item{x}{is the dataset containing the results of AFC test containing these columns: subjectName, productName,descriptorName,timeName.}

\item{subjectName}{name of the column containing the subject information}

\item{productName}{name of the column containing the subject information}

\item{descriptorName}{name of the column containing the subject information}

\item{timeName}{name of the column containing the subject information}
}
\description{
Returns the dataframe with the last click only. Also indicates the time of the last click, the result (KO or OK, entered in descriptor after an underscore) and the number of clicks
}
\examples{
data(triangular)
df=keepLastOccurence(triangular,subjectName="Paneliste")
}
