% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lawlessCorrection.r
\name{lawlessCorrection}
\alias{lawlessCorrection}
\title{lawlessCorrection}
\usage{
lawlessCorrection(numericConcentrations, triangularMatrix, p_chance = 1/3)
}
\arguments{
\item{numericConcentrations}{increasing numeric concentrations (not loggued)}

\item{triangularMatrix}{a triangular matrix from getTriangularMatrix}

\item{p_chance}{probability of good answer when no perception of difference (by chance)}
}
\description{
calculates a group threshold from lawless method (2010)
}
