% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIntensityBySubject.r
\name{plotIntensityThresholdBySubject}
\alias{plotIntensityThresholdBySubject}
\title{plotIntensityThresholdBySubject}
\usage{
plotIntensityThresholdBySubject(
  resIntensity,
  subject,
  displayAFC = FALSE,
  revertX = FALSE
)
}
\arguments{
\item{resIntensity}{result of getIntensityThreshold}

\item{subject}{name of the subject to be plotted}

\item{displayAFC}{should the result of AFC be displayed.}

\item{revertX}{if TRUE the x-axis is reverted}
}
\description{
returns a plot with the intensity in ordinate and concentration in abscissa. Segments linking the consecutive concentrations are coloured according to their specificity.
}
