\name{rata}
\alias{rata}
\docType{data}
\title{rata data}
\description{
This dataset is a toy dataset containing the results of scores for 9 concentrations and 3 subjects.
}
\usage{data(rata)}
\format{
  A data frame with 27 observations on the following 7 variables directly exported from TimeSens.
  \describe{
    \item{\code{Seance}}{Numer of the session}
    \item{\code{Repetition}}{Number of the replicate}
    \item{\code{Prise}}{Number of the intake (not relevant here)}
    \item{\code{Paneliste}}{Subject Code}
    \item{\code{Produit}}{Product Code}
    \item{\code{Descripteur}}{Descriptor }
    \item{\code{Score}}{Score for this product, subject and replicate}
     }
}
