% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisScores.R
\name{analyseScores}
\alias{analyseScores}
\title{analyseSores}
\usage{
analyseScores(
  rata,
  decreasingConcentrations = c("C9", "C8", "C7", "C6", "C5", "C4", "C3", "C2", "C1"),
  subjectName = "Panéliste",
  productName = "Produit",
  descriptorName = "Descripteur",
  timeName = "Temps",
  resName = "Res",
  scoreName = "Score"
)
}
\arguments{
\item{rata}{data.frame containing columns named as productName, subjectName and 'Score' containing a score attributed to the test}

\item{decreasingConcentrations}{vector contaning the products corresponding to the decreasing concentrations}

\item{subjectName}{name of the column containing the subject information}

\item{productName}{name of the column containing the subject information}

\item{descriptorName}{name of the column containing the subject information}

\item{timeName}{name of the column containing the subject information}

\item{resName}{name of the column containing the result "OK" or "KO" for the test.}

\item{scoreName}{name of the column containing the scores in rata data.frame}
}
\description{
Returns the plots with the scores according to their concentrations
}
\examples{
data(rata)
p=analyseScores(rata,decreasingConcentrations=c("C9","C8","C7","C6","C5","C4","C3","C2","C1"),
subjectName="Paneliste",productName="Produit",scoreName="Score")
p
}
