% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisScores.R
\name{analyseScores}
\alias{analyseScores}
\title{analyseSores}
\usage{
analyseScores(
  rata,
  subject = NULL,
  decreasingConcentrations = c("C9", "C8", "C7", "C6", "C5", "C4", "C3", "C2", "C1"),
  subjectName = "Panéliste",
  productName = "Produit",
  descriptorName = "Descripteur",
  timeName = "Temps",
  resName = "Res",
  scoreName = "Score",
  triangular = NULL,
  displayAFC = FALSE,
  displayBET = TRUE,
  representationAFC = "label",
  displayIBT = TRUE,
  decreasingNumConcentrations = NULL,
  regression = FALSE,
  revertX = FALSE,
  logY = FALSE,
  minConc = 0,
  maxConc = NULL,
  y_add = NULL,
  colors_values = NULL
)
}
\arguments{
\item{rata}{data.frame containing columns named as productName, subjectName and 'Score' containing a score attributed to the test}

\item{subject}{if NULL (default), a graph of all subjects is produced. It could also contain the name of only one ore more subjects.}

\item{decreasingConcentrations}{vector containing the names of the concentrations corresponding to the decreasing concentrations}

\item{subjectName}{name of the column containing the subject information}

\item{productName}{name of the column containing the subject information}

\item{descriptorName}{name of the column containing the subject information}

\item{timeName}{name of the column containing the subject information}

\item{resName}{name of the column containing the result "OK" or "KO" for the test.}

\item{scoreName}{name of the column containing the scores in rata data.frame}

\item{triangular}{if NULL, the obtained triangular thresholds are not represented on the graph. If not NULL, contains the dataset containing the results of AFC test containing these columns: subjectName, productName,descriptorName,timeName}

\item{displayAFC}{if TRUE, the AFC result is displayed (a point when the AFC test was succeeded, a cross if it was failed )}

\item{displayBET}{if TRUE the BET threshold is calculated and represented as a x}

\item{representationAFC}{'label' or 'point'. If label, '0' or '1' is displayed. Else, 0 corresponds to a x and 1 to a o}

\item{displayIBT}{if TRUE the IBT threshold is calculated and represented as a +}

\item{decreasingNumConcentrations}{numerical values of the used concentrations}

\item{regression}{if TRUE, adds a linear regression and its statistical indicators (R2, RMSE) to the graph}

\item{revertX}{if TRUE, allows the x-axis the to be inverted}

\item{logY}{FALSE by default. If true the intensities are logged (base = )}

\item{minConc}{a number indicating the minimal concentration (not tested, but used for threshold calculations)}

\item{maxConc}{a number indicating the maximal concentration (not tested, but used for threshold calculations)}

\item{y_add}{when representationAFC=='label', the label is displayed on the ordinate of the intensity + y_add}

\item{colors_values}{a vector containing the colors to be used. Default to "black" when length(subject)==1}
}
\description{
This function returns graphical results
}
\details{
Returns the plots with the scores according to their concentrations
}
\examples{
data(rata)
p=analyseScores(rata,decreasingConcentrations=c("C9","C8","C7","C6","C5","C4","C3","C2","C1"),
subjectName="Paneliste",productName="Produit",scoreName="Score")
p
data(triangular)
p=analyseScores(rata,decreasingConcentrations=c("C9","C8","C7","C6","C5","C4","C3","C2","C1"),
subjectName="Paneliste",productName="Produit",scoreName="Score",triangular=triangular)
}
