% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisScores.R
\name{analyseScores}
\alias{analyseScores}
\title{analyseSores}
\usage{
analyseScores(
  rata,
  decreasingConcentrations = c("C9", "C8", "C7", "C6", "C5", "C4", "C3", "C2", "C1"),
  subjectName = "Panéliste",
  productName = "Produit",
  descriptorName = "Descripteur",
  timeName = "Temps",
  resName = "Res",
  scoreName = "Score",
  triangular = NULL,
  displayAFC = FALSE,
  decreasingNumConcentrations = NULL,
  regression = FALSE,
  revertX = FALSE,
  logY = FALSE
)
}
\arguments{
\item{rata}{data.frame containing columns named as productName, subjectName and 'Score' containing a score attributed to the test}

\item{decreasingConcentrations}{vector contaning the products corresponding to the decreasing concentrations}

\item{subjectName}{name of the column containing the subject information}

\item{productName}{name of the column containing the subject information}

\item{descriptorName}{name of the column containing the subject information}

\item{timeName}{name of the column containing the subject information}

\item{resName}{name of the column containing the result "OK" or "KO" for the test.}

\item{scoreName}{name of the column containing the scores in rata data.frame}

\item{triangular}{if NULL, the obtained triangular thresholds are not represented on the graph. If not NULL, contains the dataset containing the results of AFC test containing these columns: subjectName, productName,descriptorName,timeName}

\item{displayAFC}{if TRUE, the AFC result is displayed (a point when the AFC test was succeeded, a cross if it was failed )}

\item{decreasingNumConcentrations}{numerical values of the used concentrations}

\item{regression}{if TRUE, adds a linear regression and its statistical indicators (R2, RMSE) to the graph}

\item{revertX}{if TRUE, allows the x-axis the to be inverted}
}
\description{
Returns the plots with the scores according to their concentrations
}
\examples{
data(rata)
p=analyseScores(rata,decreasingConcentrations=c("C9","C8","C7","C6","C5","C4","C3","C2","C1"),
subjectName="Paneliste",productName="Produit",scoreName="Score")
p
data(triangular)
p=analyseScores(rata,decreasingConcentrations=c("C9","C8","C7","C6","C5","C4","C3","C2","C1"),
subjectName="Paneliste",productName="Produit",scoreName="Score",triangular=triangular)
}
