% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getThreshold.r
\name{getThreshold}
\alias{getThreshold}
\title{getThreshold}
\usage{
getThreshold(
  res,
  decreasingConcentrations = c("C9", "C8", "C7", "C6", "C5", "C4", "C3", "C2", "C1"),
  subjectName = "Panéliste",
  productName = "Produit",
  descriptorName = "Descripteur",
  timeName = "Temps",
  resName = "Res",
  rata = NULL,
  decreasingNumConcentrations = NULL,
  minConc = 0,
  maxConc = NULL
)
}
\arguments{
\item{res}{result of keepLastOccurrence (or dataframe)}

\item{decreasingConcentrations}{vector containing the names of the concentrations corresponding to the decreasing concentrations}

\item{subjectName}{name of the column containing the subject information}

\item{productName}{name of the column containing the subject information}

\item{descriptorName}{name of the column containing the subject information}

\item{timeName}{name of the column containing the subject information}

\item{resName}{name of the column containing the result "OK" or "KO" for the test.}

\item{rata}{data.frame containing columns named as productName, subjectName and 'Score' containing a score attributed to the test}

\item{decreasingNumConcentrations}{vector containing the numerical concentrations (logged or not)}

\item{minConc}{a number indicating the minimal concentration (not tested, but used for threshold calculations)}

\item{maxConc}{a number indicating the maximal concentration (not tested, but used for threshold calculations)}

\item{scoreName}{name of the column containing the intensities}
}
\description{
getThreshold returns the observed sensitivity thresholds of the subjects (=0 if the subject found the right sample for all concentrations, 1 if he/she found the right sample for all concentrations but the lightest one, etc.
}
\examples{
data(triangular)
res=keepLastOccurence(triangular,subjectName="Paneliste")
seuils=getThreshold(res=res,
decreasingConcentrations=c("C9","C8","C7","C6","C5","C4","C3","C2","C1"))
data(rata)
seuils=getThreshold(res=res,
decreasingConcentrations=c("C9","C8","C7","C6","C5","C4","C3","C2","C1"),rata=rata)
}
